tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/size.attributes.parse.minus.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: size.attributes.parse.minus</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>size.attributes.parse.minus</h1>
<p class="desc">Parsing of non-negative integers</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="-100" height="-100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="size.attributes.parse.minus.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Parsing of non-negative integers");
_addTest(function(canvas, ctx) {

_assertSame(canvas.width, 300, "canvas.width", "300");
_assertSame(canvas.height, 150, "canvas.height", "150");
_assertSame(canvas.getAttribute('width'), '-100', "canvas.getAttribute('width')", "'-100'");
_assertSame(canvas.getAttribute('height'), '-100', "canvas.getAttribute('height')", "'-100'");


});
</script>
